+++ common/src/main/java/earth/terrarium/hermes/api/themes/DefaultTheme.java
@@ -4,6 +4,7 @@ import com.mojang.blaze3d.vertex.PoseStack;
 import com.teamresourceful.resourcefullib.client.utils.RenderUtils;
 import com.teamresourceful.resourcefullib.common.color.ConstantColors;
 import earth.terrarium.hermes.Hermes;
+import earth.terrarium.hermes.utils.BlitUtils;
 import earth.terrarium.hermes.utils.Divisor;
 import it.unimi.dsi.fastutil.ints.IntIterator;
 import net.minecraft.client.Minecraft;
@@ -23,7 +24,7 @@ public class DefaultTheme implements Theme {
         int vOffset = hovered ? 20 : 0;
         Gui gui = Minecraft.getInstance().gui;
         gui.blit(pose, x, y, 0, vOffset, 2, 20);
-        blitRepeating(gui, pose, x + 2, y, width - 4, 20, 2, vOffset, 196, 20);
+        BlitUtils.blitRepeating(gui, pose, x + 2, y, width - 4, 20, 2, vOffset, 196, 20);
         gui.blit(pose, x + width - 2, y, 198, vOffset, 2, 20);
         if (open) {
             gui.blit(pose, x + 4, y + 6, 0, 40, 11, 7);
@@ -31,58 +32,6 @@ public class DefaultTheme implements Theme {
             gui.blit(pose, x + 7, y + 4, 11, 40, 7, 11);
         }
         Minecraft.getInstance().font.draw(pose, text, x + 20, y + 6, ConstantColors.white.getValue());
-    }
-
-    public static void blitRepeating(Gui gui, PoseStack poseStack, int i, int j, int k, int l, int m, int n, int o, int p) {
-        int q = i;
-
-        int r;
-        for(IntIterator intIterator = slices(k, o); intIterator.hasNext(); q += r) {
-            r = intIterator.nextInt();
-            int s = (o - r) / 2;
-            int t = j;
-
-            int u;
-            for(IntIterator intIterator2 = slices(l, p); intIterator2.hasNext(); t += u) {
-                u = intIterator2.nextInt();
-                int v = (p - u) / 2;
-                gui.blit(poseStack, q, t, m + s, n + v, r, u);
-            }
-        }
-
-    }
-
-    private static IntIterator slices(int i, int j) {
-        int k = Mth.positiveCeilDiv(i, j);
-        return new Divisor(i, k);
-    }
-
-    public static void blitNineSliced(Gui gui, PoseStack poseStack, int i, int j, int k, int l, int m, int n, int o, int p, int q, int r, int s, int t) {
-        m = Math.min(m, k / 2);
-        o = Math.min(o, k / 2);
-        n = Math.min(n, l / 2);
-        p = Math.min(p, l / 2);
-        if (k == q && l == r) {
-            gui.blit(poseStack, i, j, s, t, k, l);
-        } else if (l == r) {
-            gui.blit(poseStack, i, j, s, t, m, l);
-            blitRepeating(gui, poseStack, i + m, j, k - o - m, l, s + m, t, q - o - m, r);
-            gui.blit(poseStack, i + k - o, j, s + q - o, t, o, l);
-        } else if (k == q) {
-            gui.blit(poseStack, i, j, s, t, k, n);
-            blitRepeating(gui, poseStack, i, j + n, k, l - p - n, s, t + n, q, r - p - n);
-            gui.blit(poseStack, i, j + l - p, s, t + r - p, k, p);
-        } else {
-            gui.blit(poseStack, i, j, s, t, m, n);
-            blitRepeating(gui, poseStack, i + m, j, k - o - m, n, s + m, t, q - o - m, n);
-            gui.blit(poseStack, i + k - o, j, s + q - o, t, o, n);
-            gui.blit(poseStack, i, j + l - p, s, t + r - p, m, p);
-            blitRepeating(gui, poseStack, i + m, j + l - p, k - o - m, p, s + m, t + r - p, q - o - m, p);
-            gui.blit(poseStack, i + k - o, j + l - p, s + q - o, t + r - p, o, p);
-            blitRepeating(gui, poseStack, i, j + n, m, l - p - n, s, t + n, m, r - p - n);
-            blitRepeating(gui, poseStack, i + m, j + n, k - o - m, l - p - n, s + m, t + n, q - o - m, r - p - n);
-            blitRepeating(gui, poseStack, i + k - o, j + n, m, l - p - n, s + q - o, t + n, o, r - p - n);
-        }
     }
 
     @Override
@@ -103,7 +52,7 @@ public class DefaultTheme implements Theme {
     @Override
     public void drawCraftingBackground(PoseStack pose, int x, int y, int width, int height) {
         RenderUtils.bindTexture(CRAFTING);
-        blitNineSliced(Minecraft.getInstance().gui, pose, x, y, width, height, 4, 4, 4, 4, 24, 24, 0, 0);
+        BlitUtils.blitNineSliced(Minecraft.getInstance().gui, pose, x, y, width, height, 4, 4, 4, 4, 24, 24, 0, 0);
     }
 
     @Override
@@ -111,4 +60,6 @@ public class DefaultTheme implements Theme {
         RenderUtils.bindTexture(CAROUSEL);
         Minecraft.getInstance().gui.blit(pose, x, y, left ? 0 : 14, hovered ? 20 : 0, 14, 20);
     }
+
+
 }
